+++ src/test-resources/test_inc.h
@@ -1,0 +1,8 @@
+#ifndef _TEST_INC_H_
+#define _TEST_INC_H_
+
+int included_func() {
+    return 10;
+}
+
+#endif